+++ .serverless/requirements
@@ -1,0 +1,1 @@
+/home/ezra20/.cache/serverless-python-requirements/c4f4315b14e2fdd08e041798114eaeb161885c356af35392dc3d5505d72a754a_slspyc
