+++ .serverless/requirements.txt
@@ -1,0 +1,3 @@
+feedparser==6.0.2; python_version >= "3.6"
+numpy==1.19.3; python_version >= "3.6"
+sgmllib3k==1.0.0; python_version >= "3.6"
