+++ package.json
@@ -1,0 +1,22 @@
+{
+  "name": "arxiv-sanity-preserver",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/TresTres/arxiv-sanity-preserver.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/TresTres/arxiv-sanity-preserver/issues"
+  },
+  "homepage": "https://github.com/TresTres/arxiv-sanity-preserver#readme",
+  "dependencies": {
+    "serverless-python-requirements": "^5.1.0"
+  }
+}
