+++ pyproject.toml
@@ -1,0 +1,14 @@
+[tool.poetry]
+name = "arxiv-sanity-preserver-2"
+version = "0.1.0"
+description = ""
+authors = ["Ezra Ablaza <ezra.ablaza.inbox@gmail.com>"]
+
+[tool.poetry.dependencies]
+python = "^3.8"
+numpy = "^1.19.3"
+feedparser = "^6.0.2"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
